% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{full_data}
\alias{full_data}
\title{Full data sets}
\format{
A data frame with 237,624 rows and 66 columns.
\itemize{
\item id_pe:
\item id_pe_mes:
\item no_mes:
\item year_measured:
\item essence:
\item all_cl:
\item sp_code:
\item tree_nb_sp:
\item tree_ba_sp:
\item tree_ba:
\item nb_coup:
\item recrues:
\item is_recrues:
\item is_sp_recrues:
\item is_species:
\item presence_gaule:
\item typehumus:
\item epmatorg:
\item ph_humus:
\item cl_drai:
\item soil:
\item longitude:
\item latitude:
\item altitude:
\item dom_bio:
\item logging_pr:
\item logging:
\item burn:
\item partial_logging:
\item outbreak:
\item is_logging:
\item is_partial_logging:
\item is_burn:
\item is_outbreak:
\item is_logging_pr:
\item texture:
\item year_measured_sc:
\item latitude_sc:
\item longitude_sc:
\item altitude_sc:
\item epmatorg_sc:
\item ph_humus_sc:
\item tree_ba_sc:
\item partial_logging_sc:
\item logging_sc:
\item logging_pr_sc:
\item burn_sc:
\item outbreak_sc:
\item all_cl_sc:
\item is_burn_sc:
\item is_outbreak_sc:
\item tree_ba_sp_sc:
\item id_pe_sc:
\item cl_drai_sc:
\item texture_sc:
\item an_meanT:
\item cmi_sum:
\item an_meanT_sc:
\item cmi_sum_sc:
\item have_been_species:
\item have_been_species_recrues:
\item cl_logging:
\item cl_partial_logging:
\item cl_burn:
\item cl_outbreak:
\item cl_logging_pr:
}
}
\usage{
full_data
}
\description{
Full data sets
}
\references{
\itemize{
\item https://www.donneesquebec.ca/recherche/dataset/placettes-echantillons-permanentes-1970-a-aujourd-hui
}
}
\keyword{datasets}
