% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{list_species}
\alias{list_species}
\alias{list_model}
\title{List species}
\usage{
list_species()

list_model()
}
\description{
List species
}
\section{Functions}{
\itemize{
\item \code{list_model()}: list JAGS model files

}}
