% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_jags_model}
\alias{run_jags_model}
\alias{run_analyses}
\title{Run a JAGS model}
\usage{
run_jags_model(
  sp,
  model_file,
  filename = paste0(sp, ".rds"),
  devel = FALSE,
  ...
)

run_analyses(spec, model_file, output_dir = "output", devel = FALSE, ...)
}
\arguments{
\item{sp}{species. See \code{\link[=list_species]{list_species()}}.}

\item{model_file}{JAGS model file see \code{\link[=list_model]{list_model()}}.}

\item{filename}{output filename.}

\item{devel}{a logical. If \code{TRUE}, then 2000 random rows are used instead
of the entire data set.}

\item{...}{further arguments passed to \code{\link[R2jags:jags]{R2jags::jags.parallel()}}.}

\item{spec}{A specification appropriate to the type of cluster (see
\code{\link[parallel:makeCluster]{parallel::makeCluster()}}). Beware the number of core on your computer.}

\item{output_dir}{output directory, created if missing.}
}
\description{
Function taking a species name as argument and running the jags model
it uploads the data, transform it into a list (the usable object for JAGS)
and run the model return a jags.parallel object.
}
\section{Functions}{
\itemize{
\item \code{run_analyses()}: Run analyses

}}
\examples{
\dontrun{
run_jags_model("ACERUB",  "model_time_class.bug", n.chains = 3, n.iter = 500)
}

\dontrun{
run_analyses(10, "model_time_class.bug", "output/heavy/withoutBA", 
 n.chains = 3, n.iter = 2000,)
}
}
