% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scaling}
\alias{scaling}
\title{Scaling factors}
\format{
A data frame with 30 rows and 3 columns.
\itemize{
\item var:
\item mean_sd:
\item value:
}
}
\usage{
scaling
}
\description{
Scaling factors
}
\references{
\itemize{
\item https://www.donneesquebec.ca/recherche/dataset/placettes-echantillons-permanentes-1970-a-aujourd-hui
}
}
\keyword{datasets}
